+++ package.json
@@ -27,6 +27,7 @@
     "gitalk": "^1.8.0",
     "mermaid": "^10.6.1",
     "next": "^14.0.2",
+    "next-themes": "^0.3.0",
     "notion-client": "^6.16.0",
     "notion-utils": "^6.16.0",
     "preact": "^10.19.2",
